sw_services: xilplmi: Fixes for MISRA C violation Rule 10.4
MISRA C Rule 10.4 States: Both operands of an operator in which the usual
arithmetic conversionsare performed shall have the same essential type category

This patch fixes MISRA C Rule 10.4 violation by using the same type variable
for both operands where possible


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilplmi/src/xplmi_cdo.c
@@ -404,7 +404,7 @@ int XPlmi_ProcessCdo(XPlmiCdo *CdoPtr)
 #endif
 
 	/* Verify the header for the first chunk of CDO */
-	if (CdoPtr->Cdo1stChunk == TRUE) {
+	if (CdoPtr->Cdo1stChunk == (u8)TRUE) {
 		Status = XPlmi_CdoVerifyHeader(CdoPtr);
 		if (Status != XST_SUCCESS) {
 			goto END;
@@ -432,7 +432,7 @@ int XPlmi_ProcessCdo(XPlmiCdo *CdoPtr)
 	 * In case CmdEnd is detected in previous iteration,
 	 * it just returns
 	 */
-	if (CdoPtr->CmdEndDetected == TRUE) {
+	if (CdoPtr->CmdEndDetected == (u8)TRUE) {
 		Status = XST_SUCCESS;
 		goto END;
 	}
@@ -468,7 +468,7 @@ int XPlmi_ProcessCdo(XPlmiCdo *CdoPtr)
 		 * exit the loop
 		 */
 		if ((Status != XST_SUCCESS) ||
-			(CdoPtr->CmdEndDetected == TRUE)) {
+			(CdoPtr->CmdEndDetected == (u8)TRUE)) {
 			goto END;
 		}
